+++ frontend/src/components/donor/DonorLocationPage/DonorLocationPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState }from "react";
 import { useNavigate } from "react-router-dom";
-import DonatorNavbar from "../DonatorNavbar/DonatorNavbar";
-require("./DonatorLocationPage.css");
+import DonatorNavbar from "../DonorNavbar/DonorNavbar";
+require("./DonorLocationPage.css");
 
 const DonatorLocationPage = (): JSX.Element => {
     const [address, setAddress] = useState<string>("");
