+++ frontend/src/components/donor/DonorNavbar/DonorNavbar.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import logo from "images/logo.png";
-require('./DonatorNavbar.css')
+require('./DonorNavbar.css')
 
 const navBarHeaders: string[] = [
     "Make a Donation",
