+++ frontend/src/components/donor/DonorScheduleDropoffPickupPage/DonorSchedulePickUp.tsx
@@ -6,7 +6,7 @@ import moment from "moment";
 import Checkbox from "@mui/material/Checkbox";
 import RadioButtonUncheckedIcon from '@mui/icons-material/RadioButtonUnchecked';
 import CheckCircleIcon from '@mui/icons-material/CheckCircle';
-require("./DonatorSchedulePickUp.css");
+require("./DonorSchedulePickUp.css");
 
 const weekdays = [
     "Sunday",
