+++ frontend/src/components/donor/donation/SubmitDropOffPage/SubmitDropOffPage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import DonatorNavbar from '../../DonatorNavbar/DonatorNavbar';
+import DonatorNavbar from '../../DonorNavbar/DonorNavbar';
 import SubmitInfo from "../SubmitInfo";
 import sofa1 from "./../images/sofa-01.png";
 require("./SubmitDropOffPage.css");
